refactor(domain): turn transaction into an immutable value type (#385)
Merge `transaction_basis` into `transaction` as a single value type.
Drop all cache fields and their mutexes: `hash_`, `outputs_hash_`,
`inpoints_hash_`, `sequences_hash_`, `utxos_hash_`, the optional
`total_input_value_` / `total_output_value_` / `segregated_` caches,
plus `hash_mutex_` and `mutex_`. Same philosophy as the header,
script, output and input refactors.

`hash()`, `outputs_hash()`, `inpoints_hash()`, `sequences_hash()` and
`utxos_hash()` now recompute on every call. `total_input_value()` and
`total_output_value()` recompute too. The `recompute_hash()` member
went away with the cache — the C-API test that called it via
`kth_chain_transaction_recompute_hash` was updated.

Net diff: -700+ LOC after removing `transaction_basis.hpp/.cpp` and
the per-cache lock dances.

The inner `validation` struct was renamed to `validation_t` (same
collision with the member as in output/header refactors); no public
callers referenced it by type.
